gbq_data_refresh/requirements.txt
Flask==1.1.2
gunicorn==20.0.4
google-cloud-logging==1.15.1
pandas == 0.24.2
google-cloud-bigquery==1.26.1
sodapy==2.1.0